Remove implicit conversion from factories to BuildFrom
The purpose of BuildFrom is to reuse a factory from a given source
collection. When we explicitly pass a factory in place of a BuildFrom
instance, we really want that factory to be used (not the one from
a source collection). Because the intentions behind the use of BuildFrom
and the use of an explicit factory are quite different, I think it’s
better to not have a transparent implicit conversion.